refactor: move schema types for faqs, features, hero, and text-and-image
