+++ src/sanity/schema-types/faqs-type.ts
