+++ src/sanity/schema-types/image-field-type.ts
